Remove `Buildable` instances from public API.
Some of these instances are still used for testing purposes, so relocate
them to the `Test.Utilities` module for now.
